requirements.txt
python-telegram-bot
ollama
requests
pandas
numpy
yfinance
duckduckgo-search
google-genai
scikit-learn
pyautogui
pillow